requirements.txt
psycopg2==2.9.11
python-dotenv==1.1.1
python_dateutil==2.9.0.post0
selenium==4.37.0
